その他関数を切り出し
